Implement crosschain checks
