+++ pyproject.toml
@@ -26,6 +26,7 @@ click = "^8.1.3"
 pyyaml = "^6.0"
 types-pyyaml = "^6.0.12"
 types-pytz = "^2022.4.0.0"
+more-itertools = "^9.0.0"
 
 
 [tool.poetry.group.dev.dependencies]
@@ -41,6 +42,7 @@ poethepoet = "^0.16.4"
 isort = "^5.10.1"
 black = "^22.10.0"
 mypy = "^0.982"
+ipdb = "^0.13.9"
 
 [tool.poetry.scripts]
 pyth-observer = "pyth_observer.cli:run"
