+++ pyth_observer/__init__.py
@@ -1,9 +1,8 @@
 import asyncio
-import json
 import os
 from typing import Any, Dict, Tuple
 
-import yaml
+from base58 import b58decode
 from loguru import logger
 from pythclient.pythclient import PythClient
 from pythclient.solana import (
@@ -16,11 +15,11 @@ from pythclient.solana import (
 )
 from throttler import Throttler
 
-from pyth_observer.check import PriceFeedState, PublisherState
+from pyth_observer.checks.price_feed import CrosschainPrice, PriceFeedState
+from pyth_observer.checks.publisher import PublisherState
 from pyth_observer.coingecko import get_coingecko_prices
+from pyth_observer.crosschain import CrosschainPriceObserver as Crosschain
 from pyth_observer.dispatch import Dispatch
-
-from .dns import get_key  # noqa
 
 PYTHTEST_HTTP_ENDPOINT = "https://api.pythtest.pyth.network/"
 PYTHTEST_WS_ENDPOINT = "wss://api.pythtest.pyth.network/"
@@ -47,83 +46,106 @@ def get_solana_urls(network) -> Tuple[str, str]:
 class Observer:
     def __init__(self, config: Any, publishers: Any, coingecko_mapping: Any):
         self.config = config
+        self.dispatch = Dispatch(config, publishers)
         self.publishers = publishers
+        self.pyth_client = PythClient(
+            solana_endpoint=config["network"]["http_endpoint"],
+            solana_ws_endpoint=config["network"]["ws_endpoint"],
+            first_mapping_account_key=config["network"]["first_mapping"],
+        )
+        self.pyth_throttler = Throttler(
+            rate_limit=int(config["network"]["request_rate_limit"]),
+            period=float(config["network"]["request_rate_period"]),
+        )
+        self.crosschain = Crosschain(self.config["network"]["crosschain_endpoint"])
+        self.crosschain_throttler = Throttler(rate_limit=1, period=1)
         self.coingecko_mapping = coingecko_mapping
 
     async def run(self):
-        async with PythClient(
-            solana_endpoint=self.config["network"]["http_endpoint"],
-            solana_ws_endpoint=self.config["network"]["ws_endpoint"],
-            first_mapping_account_key=self.config["network"]["first_mapping"],
-        ) as pyth:
-            pyth_rate_limit = Throttler(rate_limit=10, period=1)
-            coingecko_rate_limit = Throttler(rate_limit=1, period=60)
-            dispatch = Dispatch(self.config, self.publishers)
+        while True:
+            logger.info("Running checks")
 
-            while True:
-                logger.info("Running checks")
+            products = await self.get_pyth_products()
+            coingecko_prices, coingecko_updates = await self.get_coingecko_prices()
+            crosschain_prices = await self.get_crosschain_prices()
 
-                async with pyth_rate_limit:
-                    await pyth.refresh_all_prices()
+            for product in products:
+                # Skip tombstone accounts with blank metadata
+                if "base" not in product.attrs:
+                    continue
 
-                async with pyth_rate_limit:
-                    products = await pyth.get_products()
+                # For each product, we build a list of price feed states (one
+                # for each price account) and a list of publisher states (one
+                # for each publisher).
+                states = []
+                price_accounts = await self.get_pyth_prices(product)
+                crosschain_price = crosschain_prices.get(
+                    b58decode(product.first_price_account_key.key).hex()
+                )
 
-                async with coingecko_rate_limit:
-                    data = await get_coingecko_prices(self.coingecko_mapping)
-                    coingecko_prices: Dict[str, float] = {}
-                    coingecko_updates: Dict[str, int] = {}  # Unix timestamps
+                for _, price_account in price_accounts.items():
+                    states.append(
+                        PriceFeedState(
+                            symbol=product.attrs["symbol"],
+                            asset_type=product.attrs["asset_type"],
+                            public_key=price_account.key,
+                            status=price_account.aggregate_price_status,
+                            slot=price_account.valid_slot,
+                            slot_aggregate=price_account.aggregate_price_info.pub_slot,
+                            price_aggregate=price_account.aggregate_price_info.price,
+                            confidence_interval_aggregate=price_account.aggregate_price_info.confidence_interval,
+                            coingecko_price=coingecko_prices.get(product.attrs["base"]),
+                            coingecko_update=coingecko_updates.get(
+                                product.attrs["base"]
+                            ),
+                            crosschain_price=crosschain_price,
+                        )
+                    )
 
-                    for symbol in data:
-                        coingecko_prices[symbol] = data[symbol]["usd"]
-                        coingecko_updates[symbol] = data[symbol]["last_updated_at"]
-
-                for product in products:
-                    # Skip tombstone accounts with blank metadata
-                    if "base" not in product.attrs:
-                        continue
-
-                    async with pyth_rate_limit:
-                        price_accounts = await product.get_prices()
-
-                    states = []
-
-                    for _, price_account in price_accounts.items():
+                    for component in price_account.price_components:
                         states.append(
-                            PriceFeedState(
+                            PublisherState(
                                 symbol=product.attrs["symbol"],
-                                asset_type=product.attrs["asset_type"],
-                                public_key=price_account.key,
-                                status=price_account.aggregate_price_status,
-                                slot=price_account.valid_slot,
-                                slot_aggregate=price_account.aggregate_price_info.pub_slot,
+                                public_key=component.publisher_key,
+                                confidence_interval=component.latest_price_info.confidence_interval,
+                                confidence_interval_aggregate=component.last_aggregate_price_info.confidence_interval,
+                                price=component.latest_price_info.price,
                                 price_aggregate=price_account.aggregate_price_info.price,
-                                confidence_interval_aggregate=price_account.aggregate_price_info.confidence_interval,
-                                coingecko_price=coingecko_prices.get(
-                                    product.attrs["base"]
-                                ),
-                                coingecko_update=coingecko_updates.get(
-                                    product.attrs["base"]
-                                ),
+                                slot=component.latest_price_info.pub_slot,
+                                slot_aggregate=component.last_aggregate_price_info.pub_slot,
+                                status=component.latest_price_info.price_status,
                             )
                         )
 
-                        for component in price_account.price_components:
-                            states.append(
-                                PublisherState(
-                                    symbol=product.attrs["symbol"],
-                                    public_key=component.publisher_key,
-                                    confidence_interval=component.latest_price_info.confidence_interval,
-                                    confidence_interval_aggregate=component.last_aggregate_price_info.confidence_interval,
-                                    price=component.latest_price_info.price,
-                                    price_aggregate=price_account.aggregate_price_info.price,
-                                    slot=component.latest_price_info.pub_slot,
-                                    slot_aggregate=component.last_aggregate_price_info.pub_slot,
-                                    status=component.latest_price_info.price_status,
-                                )
-                            )
+                self.dispatch.run(states)
 
-                    dispatch.run(states)
+            logger.debug("Sleeping...")
+            await asyncio.sleep(5)
 
-                logger.debug("Sleeping for 10 seconds")
-                await asyncio.sleep(10)
+    async def get_pyth_products(self):
+        logger.debug("Fetching Pyth product accounts...")
+
+        async with self.pyth_throttler:
+            return await self.pyth_client.get_products()
+
+    async def get_pyth_prices(self, product):
+        logger.debug("Fetching Pyth price accounts...")
+
+        async with self.pyth_throttler:
+            return await product.get_prices()
+
+    async def get_coingecko_prices(self):
+        logger.debug("Fetching CoinGecko prices...")
+
+        data = await get_coingecko_prices(self.coingecko_mapping)
+        prices: Dict[str, float] = {}
+        updates: Dict[str, int] = {}  # Unix timestamps
+
+        for symbol in data:
+            prices[symbol] = data[symbol]["usd"]
+            updates[symbol] = data[symbol]["last_updated_at"]
+
+        return (prices, updates)
+
+    async def get_crosschain_prices(self) -> Dict[str, CrosschainPrice]:
+        return await self.crosschain.get_crosschain_prices()
