+++ pyth_observer/crosschain.py
@@ -1,5 +1,7 @@
 import requests
+from aiohttp import ClientSession
 from loguru import logger
+from more_itertools import chunked
 from throttler import throttle
 
 
@@ -17,24 +19,32 @@ class CrosschainPriceObserver:
             logger.error("failed to connect to cross-chain api")
             return False
 
-    # We have an attester that sends the prices to wormhole to be verified.
-    # Currently it sends prices to wormhole once every minute due to costs.
-    # We can increase the frequency of the attester once we're live on pythnet.
-    @throttle(rate_limit=1, period=60)
+    @throttle(rate_limit=1, period=1)
     async def get_crosschain_prices(self):
-        price_feed_ids = requests.get(
-            f"{self.url}/api/price_feed_ids"
-        ).json()
-        query_params = "ids[]=" + "&ids[]=".join(price_feed_ids)
-        latest_price_feeds = requests.get(
-            f"{self.url}/api/latest_price_feeds?{query_params}"
-        ).json()
-        # return a dictionary of id -> {price, conf, expo} for fast lookup
+        async with ClientSession(
+            headers={"content-type": "application/json"}
+        ) as session:
+            price_feed_ids_url = f"{self.url}/api/price_feed_ids"
+
+            async with session.get(price_feed_ids_url) as response:
+                price_feed_ids = await response.json()
+
+            price_feeds = []
+
+            for ids in chunked(price_feed_ids, 25):
+                # FIXME: Properly build this URL
+                query_params = "ids[]=" + "&ids[]=".join(ids)
+                price_feeds_url = f"{self.url}/api/latest_price_feeds?{query_params}"
+
+                async with session.get(price_feeds_url) as response:
+                    price_feeds += await response.json()
+
+        # Return a dictionary of id -> {price, conf, expo} for fast lookup
         return {
             data["id"]: {
                 "price": int(data["price"]["price"]) * 10 ** data["price"]["expo"],
                 "conf": int(data["price"]["conf"]) * 10 ** data["price"]["expo"],
                 "publish_time": data["price"]["publish_time"],
             }
-            for data in latest_price_feeds
+            for data in price_feeds
         }
